+++ datacontract/imports/importer_factory.py
@@ -104,3 +104,8 @@ importer_factory.register_lazy_importer(
     module_path="datacontract.imports.parquet_importer",
     class_name="ParquetImporter",
 )
+importer_factory.register_lazy_importer(
+    name=ImportFormat.protobuf,
+    module_path="datacontract.imports.protobuf_importer",
+    class_name="ProtoBufImporter",
+)